Remove getProperty API
